Separated server from client component and hid edits from regular users.
